LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-first-textarea-001.tentative-expected.txt








FAIL .target > * 1 assert_equals:
<span data-offset-y="29"></span>
offsetTop expected 29 but got 80
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<span data-offset-y="29"></span>
offsetTop expected 29 but got 80
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<span data-offset-y="29"></span>
offsetTop expected 29 but got 80
PASS .target > * 6
PASS .target > * 7
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
